Add Solarized theme
Adds a theme with Solarized colours. None of the other themes looked good on a
Solarized terminal. Both the old and new-fw-included versions are updated.

Due to the RGB escape sequences, an adjustment factor is added; corr3. This
makes the theme menu display with example items aligned.

Inserting the theme in the middle of the list caused other changes:
- ncorr adjusted to account for new numbering order
- themes manually renumbered where required

References:

Solarized: https://ethanschoonover.com/solarized/
Terminal colouring explanation:
    https://stackoverflow.com/a/33206814/1352761