Release amtm 6.1.2
Changed armv7l router packport repository to garycnew only. The maurerr repo is no longer used.
If you have this enabled in ep, 6, disable it and re-enable it again to only use @garycnew repo after updating to this latest amtm version.

Some changes in the mount-entware.mod as suggested by GitHub user ran-dall. This adds a cleanup logic to remove /tmp/opt if it’s a broken symlink or directory.

Good news for AdGuardHome users on firmware 3006, the woes are over and amtm caught up to its recent updates. You may again use amtm to check for updates.